Adding a mindmap
This is just a very rudimentary way of getting d3 and React together. I
will revision this soon. It's just to get something working at first.
It looks good but does not resize to the browser.
Also, I found that when adding to my Nodes - the subNodes were not
updating. This is something that needs to be fixed.
Also, changes I make in the nodes.js are not being reflected in Github.
Maybe this has to do with the .gitIgnore file. I'll give it a look when
I am in limerick.