Change Size Of MindMap to ViewPort
Checks the size of the window and sets the size in a new reducer called
viewPort. Mindmap gets the size from the reducer and sends it into
graph. This took ages to fix, because window was not something that
could be accessed inside a React component. And in the app component,
this.props.dispatch was not a function. Instead , I had to use
store.dispatch, but I did not realise for a while