First Phase of trying to make the get rid of ids
- I have decided to make the depth of the nodes unlimited.
- This means getting rid of passing parent and child ids, stopping
finding nodes by parentIndex and childIndex etc
- This is a huge change to the project. This will have massive
implications to the API. Currently I am finding things by parent and
child ids and indexes everywhere. This needs to change.
- It will clean up the code, make it more recursive. But unless I pass
an array of indexes to the API, which I am not a fan of then there will
be an exponential increase in the amout of calls looking through EVERY
node in the DB for a single ID. I dont know if this is best practice or
if I should pass in the array